+++ frontend/components/navbar/Logo.tsx
@@ -4,7 +4,7 @@ import LogoImage from "./../../public/icons/Icon_black_square.png";
 
 export default function Logo() {
   return (
-    <Link href="/dashboard" className="flex items-center mr-12">
+    <Link href="/dashboard" className="flex items-center">
       <Image priority={true} src={LogoImage} alt="Logo" width={0} height={0} className="w-12" />
     </Link>
   );
