+++ frontend/components/navbar/Navbar.tsx
@@ -8,7 +8,9 @@ export default function Navbar() {
   return (
     <div className="bg-white fixed top-0 w-screen z-50">
       <div className="flex items-center py-2 px-5 lg:px-8 xl:px-12">
-        <Logo />
+        <div className="mr-12">
+          <Logo />
+        </div>
         <SearchBar />
         <ProfilePic />
       </div>
